Refactor decode_channel_map() to lib.channel_map
This is so that the decode_channel_map() can be used for
multiple helper scripts.
